✨ Jules05: Product coherence improvements
- Centralized core enums (SignalDirection, DecisionStatus, EventImpact, EventCategory) in src/core/constants.py to eliminate redundancy and strengthen module boundaries.
- Refactored schemas, models, explainability, and event intelligence modules to use the unified types.
- Standardized terminology to "Signal" for model outputs across main.py, Makefile, and public-facing documentation.
- Created docs/quality/PRODUCT_COHERENCE_AUDIT.md to document the product's architectural health.
- Fixed minor syntax and alignment issues in main.py.
- Verified all changes with a comprehensive suite of 60+ tests.
