+++ docs/quality/PRODUCT_COHERENCE_AUDIT.md
@@ -1,44 +1,35 @@
-# Product Coherence Audit - April 2026
+# Product Coherence Audit - May 2026
 
-## 1. Product Capability
-**Status: 🟢 Foundational**
-- **Current:** Single-symbol (XAUUSD) live trading loop, MetaTrader 5 integration, basic order execution, and trade logging.
-- **Gaps:**
-    - Lack of multi-symbol portfolio management despite "All-Weather" references in `RiskManager`.
-    - No news/event filtering logic implemented in the main loop.
-    - Limited order management (simple TP/SL based on ATR, no trailing stops or partial closes in `main.py`).
+## 1. Executive Summary
+This audit evaluates the MT5 AI XAUUSD Trader for architectural coherence, naming consistency, and institutional polish. While the system demonstrates high technical maturity, several areas of "fragmentation debt" have been identified that could reduce operator trust and increase maintenance overhead.
 
-## 2. Usability
-**Status: 🟡 Minimal**
-- **Current:** CLI-driven entrypoint (`main.py`) with basic configuration via Pydantic.
-- **Gaps:**
-    - No real-time dashboard for non-technical users (Prometheus/Grafana mentioned but not fully integrated into `main.py`).
-    - Deployment requires manual environment setup or basic Docker usage; lacks "one-click" cloud deployment.
-    - Logging is standard but lacks a consolidated "intelligence briefing" for the operator.
+## 2. Findings
 
-## 3. Safety
-**Status: 🟢 Robust (Theoretical) / 🟡 Basic (Implemented)**
-- **Current:** `RiskManager` with circuit breakers and Kelly sizing.
-- **Gaps:**
-    - Health gate in `main.py` is currently missing or not explicitly blocking.
-    - Lack of automated failover between MT5 and MetaAPI cloud in the implementation.
-    - No explicit handling of extreme slippage or broker-specific execution anomalies.
+### A. Naming & Type Consistency
+- **Duplicate Enums:** `SignalDirection` is defined in both `src/core/schemas.py` and `src/core/constants.py`.
+- **Mapping Layers:** Multiple layers of mapping exist between `ModelAction` (0,1,2) and `SignalDirection` (1,-1,0), creating confusion in the prediction pipeline.
+- **Terminology Drift:** The system uses "Signal", "Action", and "Decision" interchangeably for model outputs. Standardizing on "Signal" for raw model outputs and "Decision" for risk-filtered outputs is recommended.
 
-## 4. Intelligence
-**Status: 🟡 Developing**
-- **Current:** Ensemble model structure (PPO + LSTM).
-- **Gaps:**
-    - Dreamer V3 is referenced but not implemented (lazy loading fails).
-    - Model "drift" detection is purely confidence-based; lacks performance-based trigger for retraining.
-    - Explainability (XAI) is missing; the system doesn't explain *why* it took a trade beyond "direction and confidence".
+### B. Module Boundaries
+- **Core Package bloat:** `src/core` contains diverse logic (exceptions, constants, schemas, feature engineering). While centralized, some components like `FeatureEngineer` could eventually move to `src/data` if they grow further.
+- **Circular Risk:** High potential for circular dependencies when multiple core components import from each other (e.g., `schemas` importing `constants` while `constants` might need types).
 
-## 5. Market Differentiation
-**Status: 🔴 Weak**
-- **Current:** Standard technical indicators + RL.
-- **Gaps:**
-    - No gold-specific macro sensitivity (e.g., real yields, inflation expectations, central bank gold demand).
-    - Lacks institutional-grade "What-If" simulation for XAUUSD volatility spikes.
-    - Decision cockpit doesn't highlight regime changes (e.g., "Trending" vs "Mean-Reverting" gold markets).
+### C. UX & Institutional Polish
+- **CLI Terminology:** `argparse` help messages are inconsistent in their use of "algorithm" vs "algo" vs "model".
+- **Logging:** Structlog is well-integrated, but some legacy `logging.info` calls in `main.py` lack the structured context available in other modules.
 
-## Overall Maturity Score: 6.2/10
-The system is a solid technical skeleton but lacks the "Institutional Product" polish and gold-specific intelligence that would make it a market leader.
+## 3. Remediation Plan
+
+### Immediate Fixes (PR ✨ Jules05)
+1. **Consolidate Enums:** Move `SignalDirection` and `DecisionStatus` into `src/core/constants.py`.
+2. **Harmonize Schemas:** Update `src/core/schemas.py` to import from `constants.py`.
+3. **Refactor BaseModel:** Ensure all models return a `Signal` using the unified `SignalDirection`.
+4. **Main entrypoint cleanup:** Standardize terminology in CLI flags and help text.
+
+### Long-term Recommendations
+- **Domain Separation:** Evaluate moving `feature_engineering.py` to `src/data/`.
+- **TUI Dashboard:** Transition the CLI from rich-panels to a full TUI (Decision Cockpit) for better operator experience.
+
+---
+**Audit Status:** ⚠️ DEGRADED (Fragmentation debt detected)
+**Steward:** Jules05 (yxynoty)
