+++ src/core/constants.py
@@ -4,7 +4,7 @@ src/core/constants.py
 Centralized constants and enums to ensure system-wide consistency.
 """
 
-from enum import IntEnum
+from enum import Enum, IntEnum
 
 
 class SignalDirection(IntEnum):
@@ -13,6 +13,36 @@ class SignalDirection(IntEnum):
     BUY = 1
     SELL = -1
     HOLD = 0
+
+
+class DecisionStatus(str, Enum):
+    """Augmented status levels for trade execution."""
+
+    EXECUTE = "execute"
+    CAUTION = "caution"
+    BLOCKED = "blocked"
+
+
+class EventImpact(IntEnum):
+    """Normalized event impact scores for macro events."""
+
+    LOW = 1
+    MEDIUM = 2
+    HIGH = 3
+    CRITICAL = 4
+
+
+class EventCategory(Enum):
+    """Categories of macro events relevant to XAUUSD."""
+
+    CPI = "CPI"
+    NFP = "NFP"
+    FOMC = "FOMC"
+    RATES = "RATES"
+    USD = "USD"
+    USD_MACRO = "USD_MACRO"
+    GEOPOLITICAL = "GEOPOLITICAL"
+    OTHER = "OTHER"
 
 
 class ModelAction(IntEnum):
