+++ src/core/schemas.py
@@ -13,16 +13,10 @@ License: MIT
 from __future__ import annotations
 
 from datetime import UTC, datetime
-from enum import IntEnum
 
 from pydantic import BaseModel, Field, field_validator, model_validator
 
-
-class SignalDirection(IntEnum):
-    """Standardized signal directions for schema enforcement."""
-    BUY = 1
-    SELL = -1
-    HOLD = 0
+from src.core.constants import SignalDirection
 
 
 class TradeSignal(BaseModel):
